requirements.txt
flask
line-bot-sdk
python-dotenv
psycopg2
psycopg2-binary
redis